Introduce testing framework for kafka-worker latencies
The test-worker allows to easily test the latencies for any worker config.

Requirements:
- change streaming_platform.type to kafka-testing
- set streaming_platform.example_message_file to file with json message to be used in testing
- if using docker-compose or k8s, make sure to mount example_message_file adequately

Testing flow:
The test-worker will start the regular kafka-worker which handles its
tasks as usual. Additionally, a second worker is started, which iterates
over all defined tasks and consecutively sends the example message for each
task to the respective input_topic. It listens to the output_topic and
times the end-to-end latency. It does this for
streaming_platform.test_iterations n times (default is 100) and produces
a {task}-latencies.log file with all latencies without outliers. It also
logs the 50, 99 and 99.9 latency percentiles.

To locally run the tests, just run `docker-compose up`.